+++ cookit/urls.py
@@ -16,4 +16,5 @@ Including another URLconf
 from django.urls import path
 
 urlpatterns = [
+    
 ]
